Moved pagination.py to common folder for resources as it involves common functionality
