+++ vantage6-server/vantage6/server/resource/organization.py
@@ -7,7 +7,7 @@ from http import HTTPStatus
 
 from vantage6.common import logger_name
 from vantage6.server import db
-from vantage6.server.resource.pagination import Pagination
+from vantage6.server.resource.common.pagination import Pagination
 from vantage6.server.permission import (
     Scope as S,
     Operation as P,
